move tests to separat RowPermissionTest file
