+++ src/test/scala/com/campudus/tableaux/api/content/RetrieveTest.scala
@@ -468,237 +468,6 @@ class RetrieveRowsTest extends TableauxTestBase {
     }
   }
 
-  @Test
-  @Ignore
-  def retrieveRowWithPermissions_authorized(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-rows": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["viewColumn", "viewCellValue", "viewTable"]
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-    val expectedJson: JsonObject = Json.obj(
-      "id" -> 1,
-      "permissions" -> Json.arr("onlyLinemaster"),
-      "values" -> Json.arr(
-        "table1row1",
-        1
-      )
-    )
-    val controller: TableauxController = createTableauxController(roleModel)
-    for {
-      _ <- createDefaultTable()
-      _ <- sendRequest("PATCH", "/tables/1/rows/1/annotations", Json.obj("permissions" -> Json.arr("onlyLinemaster")))
-
-      retrievedRow <- controller.retrieveRow(1, 1)
-    } yield {
-      assertJSONEquals(expectedJson, retrievedRow.getJson)
-    }
-  }
-
-  @Test
-  @Ignore
-  def retrieveRowWithPermissions_unauthorized(implicit c: TestContext): Unit =
-    exceptionTest("error.request.unauthorized") {
-      val roleModel = initRoleModel("""
-                                      |{
-                                      |  "view-rows": [
-                                      |    {
-                                      |      "type": "grant",
-                                      |      "action": ["viewColumn", "viewCellValue", "viewTable"]
-                                      |    }
-                                      |  ]
-                                      |}""".stripMargin)
-      val expectedJson: JsonObject = Json.obj(
-        "id" -> 1,
-        "permissions" -> Json.arr("onlyLinemaster"),
-        "values" -> Json.arr(
-          "table1row1",
-          1
-        )
-      )
-      val controller: TableauxController = createTableauxController(roleModel)
-      for {
-        _ <- createDefaultTable()
-        _ <- sendRequest("PATCH", "/tables/1/rows/1/annotations", Json.obj("permissions" -> Json.arr("onlyLinemaster")))
-
-        retrievedRow <- controller.retrieveRow(1, 1)
-      } yield {}
-    }
-
-  @Test
-  @Ignore
-  def retrieveRowsWithPermissions_authorized(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-rows": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["viewColumn", "viewCellValue", "viewTable"]
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-    val expectedJson: JsonArray = Json.arr(
-      Json.obj(
-        "id" -> 1,
-        "permissions" -> Json.arr("onlyLinemaster"),
-        "values" -> Json.arr(
-          "table1row1",
-          1
-        )
-      ),
-      Json.obj(
-        "id" -> 2,
-        "permissions" -> Json.arr(),
-        "values" -> Json.arr(
-          "table1row2",
-          2
-        )
-      )
-    )
-    val controller: TableauxController = createTableauxController(roleModel)
-    for {
-      _ <- createDefaultTable()
-      _ <- sendRequest("PATCH", "/tables/1/rows/1/annotations", Json.obj("permissions" -> Json.arr("onlyLinemaster")))
-
-      retrievedRows <- controller.retrieveRows(1, Pagination(None, None))
-    } yield {
-      assertJSONEquals(expectedJson, retrievedRows.getJson.getJsonArray("rows"))
-    }
-  }
-
-  @Test
-  @Ignore
-  def retrieveRowsWithPermissions_unauthorized(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-rows": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["viewColumn", "viewCellValue", "viewTable"]
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-    val expectedJson: JsonArray = Json.arr(
-      Json.obj(
-        "id" -> 2,
-        "permissions" -> Json.arr(),
-        "values" -> Json.arr(
-          "table1row2",
-          2
-        )
-      )
-    )
-    val controller: TableauxController = createTableauxController(roleModel)
-    for {
-      _ <- createDefaultTable()
-      _ <- sendRequest("PATCH", "/tables/1/rows/1/annotations", Json.obj("permissions" -> Json.arr("onlyLinemaster")))
-
-      retrievedRows <- controller.retrieveRows(1, Pagination(None, None))
-    } yield {
-      assertJSONEquals(expectedJson, retrievedRows.getJson.getJsonArray("rows"))
-    }
-  }
-
-  @Test
-  @Ignore
-  def retrieveForeignRowsWithPermissions_authorized(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-rows": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["viewColumn", "viewCellValue", "viewTable"]
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-    val expectedJson: JsonArray = Json.arr(
-      Json.obj(
-        "id" -> 1,
-        "permissions" -> Json.arr("onlyLinemaster"),
-        "values" -> Json.arr(
-          "table1row1"
-        )
-      ),
-      Json.obj(
-        "id" -> 2,
-        "permissions" -> Json.arr(),
-        "values" -> Json.arr(
-          "table1row2"
-        )
-      )
-    )
-    val createLinkColumn = Json.obj(
-      "columns" -> Json.arr(
-        Json.obj(
-          "name" -> "link col",
-          "toTable" -> 1,
-          "kind" -> "link"
-        )
-      )
-    )
-    val controller: TableauxController = createTableauxController(roleModel)
-    for {
-      _ <- createDefaultTable()
-      _ <- createDefaultTable(" link test table", 2)
-      _ <- sendRequest("PATCH", "/tables/1/rows/1/annotations", Json.obj("permissions" -> Json.arr("onlyLinemaster")))
-      _ <- sendRequest("POST", "/tables/2/columns", createLinkColumn)
-      _ <- sendRequest("PATCH", "/tables/2/columns/3/rows/1", Json.obj("value" -> 1))
-      _ <- sendRequest("PATCH", "/tables/2/columns/3/rows/1", Json.obj("value" -> 2))
-      cell <- sendRequest("GET", "/tables/2/columns/3/rows/1")
-      retrievedRows <- controller.retrieveForeignRows(2, 3, 1, Pagination(None, None))
-    } yield {
-      assertEquals(expectedJson, retrievedRows.getJson.getJsonArray("rows"))
-    }
-  }
-
-  @Test
-  @Ignore
-  def retrieveForeignRowsWithPermissions_unauthorized(implicit c: TestContext): Unit = okTest {
-    val roleModel = initRoleModel("""
-                                    |{
-                                    |  "view-rows": [
-                                    |    {
-                                    |      "type": "grant",
-                                    |      "action": ["viewColumn", "viewCellValue", "viewTable"]
-                                    |    }
-                                    |  ]
-                                    |}""".stripMargin)
-    val expectedJson: JsonArray = Json.arr(
-      Json.obj(
-        "id" -> 2,
-        "permissions" -> Json.arr(),
-        "values" -> Json.arr(
-          "table1row2"
-        )
-      )
-    )
-    val createLinkColumn = Json.obj(
-      "columns" -> Json.arr(
-        Json.obj(
-          "name" -> "link col",
-          "toTable" -> 1,
-          "kind" -> "link"
-        )
-      )
-    )
-    val controller: TableauxController = createTableauxController(roleModel)
-    for {
-      _ <- createDefaultTable()
-      _ <- createDefaultTable(" link test table", 2)
-      _ <- sendRequest("PATCH", "/tables/1/rows/1/annotations", Json.obj("permissions" -> Json.arr("onlyLinemaster")))
-      _ <- sendRequest("POST", "/tables/2/columns", createLinkColumn)
-      _ <- sendRequest("PATCH", "/tables/2/columns/3/rows/1", Json.obj("value" -> 1))
-      _ <- sendRequest("PATCH", "/tables/2/columns/3/rows/1", Json.obj("value" -> 2))
-      cell <- sendRequest("GET", "/tables/2/columns/3/rows/1")
-      retrievedRows <- controller.retrieveForeignRows(2, 3, 1, Pagination(None, None))
-    } yield {
-      assertEquals(expectedJson, retrievedRows.getJson.getJsonArray("rows"))
-    }
-  }
 }
 
 @RunWith(classOf[VertxUnitRunner])
