+++ src/test/scala/com/campudus/tableaux/api/content/RowLevelAnnotationsTest.scala
@@ -54,51 +54,3 @@ class RowLevelAnnotationsTest extends TableauxTestBase {
     }
   }
 }
-
-@RunWith(classOf[VertxUnitRunner])
-class RowPermissionAnnotationsTest extends TableauxTestBase {
-
-  @Test
-  @Ignore
-  def createRowAndSetPermissionFlagsForSingleRow(implicit c: TestContext): Unit = {
-    okTest {
-      for {
-        tableId <- createEmptyDefaultTable()
-
-        // empty row
-        result <- sendRequest("POST", s"/tables/$tableId/rows")
-        rowId = result.getLong("id")
-
-        _ <- sendRequest(
-          "PATCH",
-          s"/tables/$tableId/rows/$rowId/annotations",
-          Json.obj("permissions" -> Json.arr("perm_1", "perm_2"))
-        )
-
-        rowJson1 <- sendRequest("GET", s"/tables/$tableId/rows/$rowId")
-
-        _ <- sendRequest("PATCH", s"/tables/$tableId/rows/$rowId/annotations", Json.obj("permissions" -> Json.arr()))
-
-        rowJson2 <- sendRequest("GET", s"/tables/$tableId/rows/$rowId")
-      } yield {
-        val expectedRowJson1 = Json.obj(
-          "status" -> "ok",
-          "id" -> rowId,
-          "permissions" -> Json.arr("perm_1", "perm_2"),
-          "values" -> Json.arr(null, null)
-        )
-
-        assertJSONEquals(expectedRowJson1, rowJson1)
-
-        val expectedRowJson2 = Json.obj(
-          "status" -> "ok",
-          "id" -> rowId,
-          "permissions" -> Json.arr(),
-          "values" -> Json.arr(null, null)
-        )
-
-        assertJSONEquals(expectedRowJson2, rowJson2)
-      }
-    }
-  }
-}
